Make domain configurable
So far the uberspace installer always made the service available at the
standard domain for the account. This meant that it couldn't be used
with a custom domain registered for the uberspace. In addition, it
didn't play nice with an existing service at the default document root,
potentially overwriting an existing `.htaccess` file.

This commit allows the user to specify a custom domain (including
subdomains) and the corresponding document root.